♻️ refactor: 카테고리 리스트 조회 API 요청 코드 분리 리팩토링
